+++ src/app/controller/userController.ts
@@ -4,6 +4,7 @@ import mongoose from 'mongoose';
 import User from '../model/user'
 import signJWT from '../../function/singJWT'
 import bcrypt from "bcrypt";
+import Category from '../model/categories'
 class UserController {
     public async create(req: Request, res: Response){
         const {username, password, email} = req.body
@@ -53,6 +54,35 @@ class UserController {
             }
         });
 
+    }
+
+    public async update(req: Request, res: Response){
+        const body = req.body;
+        const userId = res.locals.jwt.id
+        const user = await User.findOne({ _id: userId });
+
+        if(!user) return res.status(404).json({ message: 'user not found'})
+
+        user!.username = body.username;
+        await user.save()
+
+        return res.status(200).send(user)
+    }
+
+    public async delete(req: Request, res: Response){
+        const userId = res.locals.jwt.id
+        const user = await User.findOne({ _id: userId });
+
+        if(!user) return res.status(404).json({ message: 'user not found'})
+
+        try{
+            await Category.deleteMany({user_id: userId})
+            await user.delete()
+
+            return res.status(200).send({message: `the user ${user.username} deleted`})
+        }catch(err){
+            return res.status(500).json({ message:'somenthing went wrong'})
+        }        
     }
 }