+++ src/routes/routes.ts
@@ -14,6 +14,8 @@ router.get('/ping', (req: Request ,res: Response) => {
  */
 router.post('/create_user',UserController.create)
 router.post('/login_user', UserController.login)
+router.put('/update_user', extractJWT , UserController.update)
+router.delete('/delete_user', extractJWT , UserController.delete)
 
 /**
  * middleware
@@ -26,6 +28,7 @@ router.use(extractJWT)
 router.post('/create_category', CategoryController.createCategory)
 router.delete('/delete_category/:category', CategoryController.deleteCategory)
 router.get('/get_categories' ,CategoryController.getCategories)
+router.put('/get_categories' ,CategoryController.update)
 
 /**
  * tags routes
@@ -33,5 +36,6 @@ router.get('/get_categories' ,CategoryController.getCategories)
 router.post('/create_tag', TagsController.create)
 router.delete('/delete_tag/:tag/:category', TagsController.delete)
 router.get('/get_tags/:category', TagsController.tagsList)
+router.put('/get_tags/:category', TagsController.update)
 
 export default router