+++ changelog.md
@@ -5,8 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.0] - 2017-06-20
+### Added
+- controller and routes to tags
+- controller delete and update user
+- tag and category update method
+## [Unreleased]
 * tags controller
-- change
+* category controller
+- update
 
 ## [0.1.0] - 2017-06-20
 ### Added